requirements.txt
Flask
flask-mongoengine
line-bot-sdk
openai
python-dotenv
